% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{adj_to_edgelist}
\alias{adj_to_edgelist}
\title{Convert representation of time-sliced network as N x N x T array to a length-T list of edglists for each time period. (Handle NA?)
Resulting edgelists have three columns, "from", "to", and "count" (which is more generally the edge weight).}
\usage{
adj_to_edgelist(A, directed = FALSE, selfedges = TRUE)
}
\arguments{
\item{A}{is a (time) series of network data represented as a N x N x Time array (each slice represented as an adjacency matrix).}

\item{directed}{Are the edges in the edgelist directed?}

\item{selfedges}{If true, include self-edges in output edgelists. If false, remove. Note tdsbm methods allow/include selfedges.}
}
\description{
Convert representation of time-sliced network as N x N x T array to a length-T list of edglists for each time period. (Handle NA?)
Resulting edgelists have three columns, "from", "to", and "count" (which is more generally the edge weight).
}
