% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sbmt.R
\name{plot.sbmt}
\alias{plot.sbmt}
\title{generic function to plot block-to-block edges (i.e. sum of edges of all nodes in the block) over time (reverse direction is show in )}
\usage{
\method{plot}{sbmt}(
  x,
  mai = rep(0.5, 4),
  show_blocks = list(1:nrow(x$EdgeMatrix[[1]]), 1:nrow(x$EdgeMatrix[[1]])),
  show_reverse = TRUE,
  ...
)
}
\arguments{
\item{x}{output of sbmt of class sbmt}

\item{mai}{sets plot margin width}

\item{show_blocks}{is list of length two that sets which block-to-block edges to show (defaults to all)}

\item{show_reverse}{if true will show the reverse direction as a dashed gray line (only useful for directed, but may not want even then)}

\item{...}{ignored}
}
\description{
generic function to plot block-to-block edges (i.e. sum of edges of all nodes in the block) over time (reverse direction is show in )
}
