% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbm.R
\name{sbm}
\alias{sbm}
\title{Estimate parameters of a static SBM by Kernighan-Lin algorithm. Allows for directed networks and degree correction.}
\usage{
sbm(edgelist, maxComms = 2, degreeCorrect = F, directed = FALSE,
  klPerNetwork = 50, tolerance = 1e-04, seedComms = NULL,
  seed = NULL)
}
\description{
Estimate parameters of a static SBM by Kernighan-Lin algorithm. Allows for directed networks and degree correction.
}
