% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmt.R
\name{sbmt}
\alias{sbmt}
\title{Estimate parameters of the TDD-SBM (time-dependent discrete stochastic block model) by Kernighan-Lin algorithm.
Note that although the motivation for this function and package is analyzing time-sliced networks, it can handle multilayer networks more generally as long as the primary input is a list of networks in edgelist format. (Slice parameters are independent given block membership.)}
\usage{
sbmt(
  edgelist.time,
  maxComms = 2,
  degreeCorrect = 0,
  directed = F,
  klPerNetwork = 50,
  tolerance = 1e-04,
  seed = NULL,
  seedComms = NULL
)
}
\arguments{
\item{edgelist.time}{A (time) series of networks represented as a list of edgelists. 
Assumes all edgelist slices have the same names and number of columns. The first two columns designate edges "from" and "to", and the third, if present, is the count for each edge.}

\item{maxComms}{maximum number of communities (blocks) represented in the network}

\item{degreeCorrect}{whether to apply degree correction:
\itemize{
  \item 0 = no degree correction
  \item 1 = the degree correction of Karrer and Newman at every time step - inherits directedness of graph
  \item 2 = time-independent degree correction over time-dependent data - inherits directedness of graph
  \item 3 = time-independent degree correction over time-dependent data - undirected regardless
}}

\item{directed}{whether the network is directed and off-diagonal block parameters may be assymetrical}

\item{klPerNetwork}{this is the number of KL runs on a network}

\item{tolerance}{stopping criteria for KL. Prevents loops due to numerical errors.}

\item{seed}{a random seet set for reproducibility.}

\item{seedComms}{user-supplied starting values for KL runs. They will be converted to integer levels numbered starting at 0.
To avoid confusion, they must have names equal to the unique nodes in the network. E.g., If nodes are numbered numerically from 1 then seedComm names are "1", "2",...}
}
\value{
FoundComms A vector of node labels with estimated block assignments.

EdgeMatrix time-specific block-to-block edges corresponding to FoundComms

theta nodal degree correction parameters. NA if degreeCorrect == 0; not yet implemented for degreeCorrect == 1; 
N x 2 data frame (one column for each direction) if network is directed and degreeCorrect == 2; and a length-N numeric vector otherwise.

llik unnormalized log-likelihood of result as calculated by tdd_sbm_llik

directed Whether input network was considered directed

degreeCorrect type of degreeCorrection used in fitting

klPerNetwork number of runs of KL algorithm used in fitting

tolerance tolerance value used for fitting

init seed = set for fitting, if supplied;  seedComms = initial communities (blocks) as fed to sbmtFit (conversion to zero-min integer vector may have occurred)
}
\description{
Estimate parameters of the TDD-SBM (time-dependent discrete stochastic block model) by Kernighan-Lin algorithm.
Note that although the motivation for this function and package is analyzing time-sliced networks, it can handle multilayer networks more generally as long as the primary input is a list of networks in edgelist format. (Slice parameters are independent given block membership.)
}
