% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{tdd_sbm_llik}
\alias{tdd_sbm_llik}
\title{Calculate un-normalized log-likelihood for a TDD-SBM}
\usage{
tdd_sbm_llik(
  A,
  roles,
  omega,
  degreeCorrect = 3,
  directed = TRUE,
  selfEdges = TRUE
)
}
\arguments{
\item{A}{(time) series of network data represented  as N x N x Time array.}

\item{roles}{is a length-N list of estimated block assignment for each node. If not 0-indexed will substract min value so roles are 0,1,2,...}

\item{omega}{is a Time x K x K array describing block-to-block traffic at each time period OR Time-length list of K x K matrices.}

\item{degreeCorrect}{degree correction type:
\itemize{
  \item 0 = no degree correction
  \item 1 = the degree correction of Karrer and Newman at every time step - inherits directedness of graph
  \item 2 = time-independent degree correction over time-dependent data - inherits directedness of graph
  \item 3 = time-independent degree correction over time-dependent data - undirected regardless
}}

\item{directed}{Is the network directed?}

\item{selfEdges}{is whether to sum over self-edge indices in the likelihood calculation, or exclude them}
}
\description{
Calculate un-normalized log-likelihood for a TDD-SBM (time-dependent discrete-membership stochastic block model).
This is not the same as the "highest score" returned by sbmt, which is unnormalized in a different way
}
\examples{
data("la_byhour_edgelist")
A = edgelist_to_adj(la_byhour_edgelist, as.array = TRUE)
model1 = sbmt(la_byhour_edgelist,  degreeCorrect = 3, 
  directed = TRUE, klPerNetwork = 2, maxComms = 3)
tdd_sbm_llik(A, model1$FoundComms, model1$EdgeMatrix)
}
