% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zach}
\alias{zach}
\title{Zachary's Karate Club network (used for testing).}
\format{A data frame with 78 rows and 2 variables:
\describe{
  \item{X1}{first node involved in an edge}
  \item{X2}{second node involved in an edge}
}}
\source{
Krivitsky P (2019). _ergm.count: Fit, Simulate and Diagnose
Exponential-Family Models for Networks with Count Edges_. The Statnet
Project (<URL: https://statnet.org>). R package version 3.4.0, <URL:
}
\usage{
zach
}
\description{
The two columns list the edges by their associated nodes
}
\details{
library(ergm.count)
data(zach)
zach = data.frame(as.edgelist(zach))
save(zach, file="../sbm/data/zach.rda")
}
\references{
Zachary, WW (1977). An Information Flow Model for Conflict and Fission in Small Groups.
Journal of Anthropological Research, 33(4), 452-473.
Sociomatrix in machine-readable format was retrieved from http://vlado.fmf.uni-lj.si/pub/

Zachary, WW (1977). An Information Flow Model for Conflict and Fission in Small Groups.
Journal of Anthropological Research, 33(4), 452-473.
}
\keyword{datasets}
