% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbmt.R
\name{sbmt}
\alias{sbmt}
\title{Estimate parameters of the TDD-SBM (time-dependent discrete stochastic block model) by Kernighan-Lin algorithm.}
\usage{
sbmt(edgelist.time, maxComms = 2, degreeCorrect = 0, directed = F,
  klPerNetwork = 50, tolerance = 1e-04, seedComms = NULL,
  seed = NULL)
}
\arguments{
\item{maxComms}{maximum number of communities represented in the network}

\item{degreeCorrect}{whether to apply degree correction.:
\itemize{
  \item 0 = no degree correction
  \item 1 = the degree correction of Karrer and Newman at every time step
  \item 2 = time-independent degree correction over time-dependent data - inherets directedness of graph
  \item 3 = time-independent degree correction over time-dependent data - undirected regardless
}}

\item{directed}{whether the network is directed and off-diagonal block parameters may be assymetrical}

\item{klPerNetwork}{this is the number of KL runs on a network}

\item{tolerance}{stopping criteria for KL. Prevents loops due to numerical errors.}

\item{seedComms}{user-supplied starting values for KL runs. NOT YET IMPLEMENTED.}

\item{seed}{a random seet set for reproducibility. (Implemented?)}
}
\value{
FoundComms is a vector of node labels with estimated block assignments.

EdgeMatrix is
}
\description{
Estimate parameters of the TDD-SBM (time-dependent discrete stochastic block model) by Kernighan-Lin algorithm.
}
