% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{td_sbm_llik_t}
\alias{td_sbm_llik_t}
\title{Calculate log-likelihood for a single timeslice given mu as input (works for discrete or continuous). Helper for tdd_sbm_llik and tdmm_sbm_llik
#}
\usage{
td_sbm_llik_t(A_t, mu, K = 2)
}
\arguments{
\item{A_t}{is a single-time network, represented as a N x N adjacency matrix}

\item{mu}{mu is a N x N matrix of edge expected values at this time slice}

\item{K}{is the number of blocks}
}
\description{
Calculate log-likelihood for a single timeslice given mu as input (works for discrete or continuous). Helper for tdd_sbm_llik and tdmm_sbm_llik
#
}
