% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_sbm_llik.R
\name{tdd_n_param}
\alias{tdd_n_param}
\title{Calculate number of parameters for discrete model}
\usage{
tdd_n_param(N, K, Time, directed = TRUE)
}
\arguments{
\item{N}{number of nodes}

\item{K}{number of blocks}

\item{Time}{number of time slices}

\item{directed}{Is the network directed?}
}
\description{
Calculate number of parameters for discrete model
}
